"Retiro de productos version 1"

+++ lib/business_logic/models/ad_model.dart
@@ -10,7 +10,7 @@ class Ad {
   final String publicationType;
   final int minimumPurchase;
   final int minimumStock;
-  final double? precioProduct;
+  final int? precioProduct;
   final String status; 
   final String emailVendedor;
   final String? refIdGroup;
